+++ src/hooks/useCurrentRef.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from 'react';
 
-export default function useCurrentRef<T>(toRef: T) {
+export function useCurrentRef<T>(toRef: T) {
   const toRefRef = useRef<T>(toRef);
   useEffect(() => {
     toRefRef.current = toRef;