+++ src/index.ts
@@ -1,1 +1,4 @@
 export * from './components';
+export * from './data';
+export * from './hooks';
+export * from './utils';